# openapi.yaml
openapi: 3.0.0
info:
  title: SportsDeck API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/appeals:
    post:
      summary: Submit a ban appeal
      tags:
        - Appeals
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  example: >-
                    I believe my ban was issued in error. I did not violate any
                    rules.
      responses:
        '201':
          description: Appeal submitted
        '400':
          description: Missing message, invalid JSON, or user is not banned
        '401':
          description: Unauthorized
        '404':
          description: No active ban found
        '409':
          description: Pending appeal already exists for this ban
    get:
      summary: Get the authenticated user's own appeals
      tags:
        - Appeals
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - pending
              - approved
              - rejected
          example: pending
        - in: query
          name: page
          schema:
            type: integer
          example: 1
        - in: query
          name: limit
          schema:
            type: integer
          example: 20
      responses:
        '200':
          description: Paginated list of user's appeals
        '400':
          description: Invalid status value
        '401':
          description: Unauthorized
...
  /api/feed/{id}/read:
    patch:
      summary: Mark a feed entry as read
      tags:
        - Feed
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: clxfeed001
      responses:
        '200':
          description: Feed entry marked as read
        '401':
          description: Unauthorized
        '404':
          description: Feed entry not found
        '500':
          description: Internal server error
  /api/feed:
    get:
      summary: Get the authenticated user's personalized activity feed
      tags:
        - Feed
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          example: 20
      responses:
        '200':
          description: List of grouped feed entries
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/follow/{userId}:
    post:
      summary: Follow a user
      tags:
        - Follow
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          example: clxuser789
      responses:
        '200':
          description: Now following the user
        '400':
          description: Cannot follow yourself or already following
        '401':
          description: Unauthorized
        '403':
          description: Account banned
        '500':
          description: Internal server error
    delete:
      summary: Unfollow a user
      tags:
        - Follow
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          example: clxuser789
      responses:
        '200':
          description: Successfully unfollowed
        '401':
          description: Unauthorized
        '403':
          description: Account banned
        '500':
          description: Internal server error
...